backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-multipart==0.0.9
aiofiles==23.2.1
pytesseract==0.3.10
Pillow==10.3.0
pdfplumber==0.11.0
PyMuPDF==1.24.3
openai==1.30.1
supabase==2.4.6
python-dotenv==1.0.1
httpx==0.27.0
